--- a/pkg/controllers/provisioning/scheduling/existingnode.go
+++ b/pkg/controllers/provisioning/scheduling/existingnode.go
@@ -89,7 +89,7 @@ func (n *ExistingNode) Add(ctx context.Context, kubeClient client.Client, pod *v
 	nodeRequirements := scheduling.NewRequirements(n.requirements.Values()...)
 	podRequirements := scheduling.NewPodRequirements(pod)
 	// Check NodeClaim Affinity Requirements
-	if err = nodeRequirements.StrictlyCompatible(podRequirements); err != nil {
+	if err = nodeRequirements.Compatible(podRequirements); err != nil {
 		return err
 	}
 	nodeRequirements.Add(podRequirements.Values()...)
@@ -106,7 +106,7 @@ func (n *ExistingNode) Add(ctx context.Context, kubeClient client.Client, pod *v
 	if err != nil {
 		return err
 	}
-	if err = nodeRequirements.StrictlyCompatible(topologyRequirements); err != nil {
+	if err = nodeRequirements.Compatible(topologyRequirements); err != nil {
 		return err
 	}
 	nodeRequirements.Add(topologyRequirements.Values()...)