<!-- Please follow the guidelines at https://www.conventionalcommits.org/en/v1.0.0/ and use one of the following in your title:
feat:            <-- New features that require a MINOR version update
fix:             <-- Bug fixes that require at PATCH version update
chore:           <-- Smaller changes that impact behavior but aren't large enough to be features
perf:            <-- Code changes that improve performance but do not impact behavior
docs:            <-- Documentation changes that do not impact code
test:            <-- Test changes that do not impact behavior
ci:              <-- Changes that affect test or rollout automation
!${type}:        <-- Include ! if your change includes a backwards incompatible change.
-->

Fixes #N/A <!-- issue number -->

**Description**

This change collapses the `Compatible` and `StrictlyCompatible` check into a single command that takes in a variadic list of options that currently allows you to specify values that can be allowed to be undefined in the requirement set. By default, we have `AllowUndefinedWellKnownLabelsV1Alpha5` and `AllowUndefinedWellKnownLabelsV1Beta1` options which allow the compatibility checks to ignore well-known labels for each API version.

**How was this change tested?**

`make presubmit`

### Before PR (Scheduling Benchmarking)

```
=== RUN   TestSchedulingProfile
scheduled 7610 against 21 nodes in total in 2.776334936s 2741.0237508893993 pods/sec
400 instances 10 pods    1 nodes  1.75316ms per scheduling     175.316µs per pod
400 instances 100 pods   1 nodes  17.786427ms per scheduling   177.864µs per pod
400 instances 500 pods   1 nodes  85.996651ms per scheduling   171.993µs per pod
400 instances 1000 pods  3 nodes  190.301ms per scheduling     190.301µs per pod
400 instances 1500 pods  4 nodes  299.967896ms per scheduling  199.978µs per pod
400 instances 2000 pods  5 nodes  412.911611ms per scheduling  206.455µs per pod
400 instances 2500 pods  6 nodes  482.912916ms per scheduling  193.165µs per pod
--- PASS: TestSchedulingProfile (13.82s)
PASS
```

### After PR (Scheduling Benchmarking)

```
=== RUN   TestSchedulingProfile
scheduled 7610 against 21 nodes in total in 2.790164606s 2727.437651397116 pods/sec
400 instances 10 pods    1 nodes  1.755514ms per scheduling    175.551µs per pod
400 instances 100 pods   1 nodes  17.707346ms per scheduling   177.073µs per pod
400 instances 500 pods   1 nodes  90.669413ms per scheduling   181.338µs per pod
400 instances 1000 pods  3 nodes  195.117416ms per scheduling  195.117µs per pod
400 instances 1500 pods  4 nodes  309.345593ms per scheduling  206.23µs per pod
400 instances 2000 pods  5 nodes  405.330403ms per scheduling  202.665µs per pod
400 instances 2500 pods  6 nodes  468.279694ms per scheduling  187.311µs per pod
--- PASS: TestSchedulingProfile (13.90s)
PASS
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.