@jonathan-innis 2023-09-06
Benchmarking just the function shows that the function runs slightly slower but this is a reasonable trade-off given code readability and simplicity

```go
func BenchmarkRequirementsOld(b *testing.B) {
	reqs := scheduling.NewRequirements()
	reqs2 := scheduling.NewRequirements()
	for i := 0; i < 100; i++ {
		reqs.Add(scheduling.NewRequirement(fmt.Sprintf("key%d", i), v1.NodeSelectorOpIn, "value"))
		reqs2.Add(scheduling.NewRequirement(fmt.Sprintf("key%d", i), v1.NodeSelectorOpIn, "value"))
	}
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		reqs.CompatibleOld(reqs2)
	}
}

func BenchmarkRequirementsNew(b *testing.B) {
	reqs := scheduling.NewRequirements()
	reqs2 := scheduling.NewRequirements()
	for i := 0; i < 100; i++ {
		reqs.Add(scheduling.NewRequirement(fmt.Sprintf("key%d", i), v1.NodeSelectorOpIn, "value"))
		reqs2.Add(scheduling.NewRequirement(fmt.Sprintf("key%d", i), v1.NodeSelectorOpIn, "value"))
	}
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		reqs.Compatible(reqs2, scheduling.AllowUndefinedWellKnownLabelsV1Alpha5)
	}
}
```

```console
BenchmarkRequirementsOld
BenchmarkRequirementsOld-10    	   12974	     92027 ns/op
PASS

BenchmarkRequirementsNew
BenchmarkRequirementsNew-10    	   12880	     93375 ns/op
PASS
```